requirements.txt
disnake
pyyaml
